scripts/requirements.txt
charset-normalizer>=3.3.2
certifi>=2023.7.22
annotated-types>=0.6.0
bech32>=1.2.0
ecdsa>=0.19.1
idna>=3.4
pydantic>=2.12.3
requests>=2.32.0
six>=1.16.0
typing_extensions>=4.12.2
typing_inspection>=0.4.2
uagents-core==0.3.11
python-dotenv>=1.0.1
fastapi>=0.115.5
uvicorn>=0.30.6
uagents>=0.12.0
openai>=1.55.0
